Финальный аудит: исправление 10 проблем надёжности и корректности
Критические исправления:
- WindowGLFW::SwapBuffers — добавлена проверка m_Window != null
- RendererOGL::Init — gladLoadGLLoader failure теперь обрабатывается в
  Release-сборке (assert — no-op в Release)
- ModuleManager::LoadModule — null-check после CreateFactory()
- RenderResourceManager::CreateTextureFromFile — инициализация w/h/ch=0,
  валидация размеров, static_cast вместо C-cast

Исправления portability:
- std::cosf() → std::cos() (std::cosf нестандартен для MSVC)
- WindowGLFW::GetRenderInitInfo — reinterpret_cast<void*> вместо
  brace-init (brace-init не выполняет function ptr → void* конверсию)

Ранее исправлено в этой сессии:
- ModuleManager: shadow copy только на Windows (macOS/Linux не блокируют dylib)
- SceneSerializer: bounds check enum LightType, валидация путей текстур
- RenderData.h: именованные struct в union (убраны 4 warning'а)
- SceneViewLayer_Panels: ImGui::PushID/PopID в цикле мешей,
  char buf[] вместо const_cast для InputText

+++ Sandbox/src/SceneViewLayer_Render.cpp
@@ -155,10 +155,10 @@ void SceneViewLayer::RenderScene()
         shader->SetFloat(CHEngine::String(nameBuf), light.Range);
 
         snprintf(nameBuf, sizeof(nameBuf), "u_LightInnerCone[%d]", i);
-        shader->SetFloat(CHEngine::String(nameBuf), cosf(glm::radians(light.InnerCone)));
+        shader->SetFloat(CHEngine::String(nameBuf), std::cos(glm::radians(light.InnerCone)));
 
         snprintf(nameBuf, sizeof(nameBuf), "u_LightOuterCone[%d]", i);
-        shader->SetFloat(CHEngine::String(nameBuf), cosf(glm::radians(light.OuterCone)));
+        shader->SetFloat(CHEngine::String(nameBuf), std::cos(glm::radians(light.OuterCone)));
     }
 
     // Если нет источников — добавляем дефолтный directional, чтобы сцена не была чёрной
